requirements.txt
pyqt5
pyyaml
erniebot
pyqt5
pyperclip
matplotlib
pyserial # for desktop pet only
